+++ src/main/java/com/api/entity/SxPayloadWeights.java
@@ -16,8 +16,8 @@ public class SxPayloadWeights {
     private String name;
 
     @JsonProperty("kg")
-    private long kg;
+    private Integer kg;
 
     @JsonProperty("lb")
-    private long lb;
+    private Integer lb;
 }
